# src/config/swagger.yaml
openapi: 3.0.0
info:
  title: Leave Management API
  version: 1.0.0
  description: MERN Leave Management Backend
servers:
  - url: http://localhost:5000/api/v1
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        lastName:
          type: string
        department:
          type: string
        position:
          type: string
        contact:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [employee, admin]
        leaveBalance:
          type: integer
          default: 20
        isTrashed:
          type: boolean
          default: false
        trashedAt:
          type: string
          format: date-time
          nullable: true
        isDeleted:
          type: boolean
          default: false
        deletedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
security:
  - bearerAuth: []
...
paths:
  /auth/login:
    post:
      summary: Login user
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: aljun.dalman@neu.edu.ph
                password:
                  type: string
                  example: Admin123
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Invalid credentials
  /auth/me:
    get:
      summary: Get current user profile
      tags: [Auth]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found
  /auth/change-password:
    post:
      summary: Change password
      tags: [Auth]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                currentPassword:
                  type: string
                  example: Admin123
                newPassword:
                  type: string
                  example: NewPass456
      responses:
        '200':
          description: Password changed successfully
        '400':
          description: Current password incorrect
...